move compatibility tests back. full class names are used in binary
