+++ checkstyle/suppressions.xml
@@ -340,8 +340,7 @@
     <suppress checks="LineLength|VisibilityModifier|AvoidStarImport|RegexpSingleline|ParameterNumber"
               files="src[\\/]main[\\/]java[\\/]com[\\/]hazelcast[\\/]internal[\\/]longregister[\\/]client[\\/]codec[\\/]"/>
     <suppress checks="" files="src[\\/]test[\\/]java[\\/]com[\\/]hazelcast[\\/]client[\\/]protocol[\\/]compatibility[\\/]"/>
-    <suppress checks=""
-              files="src[\\/]test[\\/]java[\\/]com[\\/]hazelcast[\\/]internal[\\/]serialization[\\/]impl[\\/]compatibility[\\/]"/>
+    <suppress checks="" files="src[\\/]test[\\/]java[\\/]com[\\/]hazelcast[\\/]nio[\\/]serialization[\\/]compatibility[\\/]"/>
 
     <!-- module-info.java -->
     <suppress checks="" files="[\\/]module-info"/>
